requirements.txt
python-dotenv
llama-index
openai
click